Fixing darwin
I was able to get it working on macos, sim is still not working. You need to create symbolic links for the osx version to get it to work. Plan on writing a dart function or .sh file to either download the items from github or create them. This currently breaks the windows version and it will need to be fixed.